Modified horse receiver.

+++ src/main/java/com/epam/horsebetting/receiver/impl/HorseReceiverImpl.java
@@ -3,12 +3,14 @@ package com.epam.horsebetting.receiver.impl;
 import com.epam.horsebetting.config.RequestFieldConfig;
 import com.epam.horsebetting.config.MessageConfig;
 import com.epam.horsebetting.dao.impl.HorseDAOImpl;
+import com.epam.horsebetting.database.TransactionManager;
 import com.epam.horsebetting.entity.Horse;
 import com.epam.horsebetting.exception.DAOException;
 import com.epam.horsebetting.exception.ReceiverException;
 import com.epam.horsebetting.receiver.AbstractReceiver;
 import com.epam.horsebetting.receiver.HorseReceiver;
 import com.epam.horsebetting.request.RequestContent;
+import com.epam.horsebetting.type.HorseGenderType;
 import com.epam.horsebetting.util.MessageWrapper;
 import com.epam.horsebetting.validator.HorseValidator;
 import org.apache.logging.log4j.Level;
@@ -38,19 +40,20 @@ public class HorseReceiverImpl extends AbstractReceiver implements HorseReceiver
      */
     @Override
     public void createHorse(RequestContent content) throws ReceiverException {
+        Locale locale = (Locale) content.findSessionAttribute(SESSION_LOCALE);
+
+        MessageWrapper messages = new MessageWrapper();
+        MessageConfig messageResource = new MessageConfig(locale);
+
         String name = content.findParameter(RequestFieldConfig.Horse.NAME_FIELD);
         String genderAttr = content.findParameter(RequestFieldConfig.Horse.GENDER_FIELD);
         String ageAttr = content.findParameter(RequestFieldConfig.Horse.AGE_FIELD);
         String suitAttr = content.findParameter(RequestFieldConfig.Horse.SUIT_FIELD);
 
-        Locale locale = (Locale)content.findSessionAttribute(SESSION_LOCALE);
         HorseValidator validator = new HorseValidator(locale);
 
-        MessageWrapper messages = new MessageWrapper();
-        MessageConfig messageResource = new MessageConfig(locale);
-
         if (validator.validateCreateHorse(name, genderAttr, ageAttr, suitAttr)) {
-            boolean gender = "male".equals(genderAttr);
+            boolean gender = HorseGenderType.isMale(genderAttr);
             byte age = Byte.parseByte(ageAttr);
             int suitId = Integer.parseInt(suitAttr);
 
@@ -62,14 +65,32 @@ public class HorseReceiverImpl extends AbstractReceiver implements HorseReceiver
 
             LOGGER.log(Level.DEBUG, "Want create horse: " + horse);
 
-            try (HorseDAOImpl horseDAO = new HorseDAOImpl(false)) {
-                Horse createdHorse = horseDAO.create(horse);
+            HorseDAOImpl horseDAO = new HorseDAOImpl(true);
 
-                messages.add(messageResource.get("dashboard.horse.create.success"));
-                content.insertSessionAttribute(REQUEST_MESSAGES, messages);
+            TransactionManager transaction = new TransactionManager(horseDAO);
+            transaction.beginTransaction();
 
-                LOGGER.log(Level.DEBUG, "Created horse: " + createdHorse);
+            try {
+                if (horseDAO.findByName(name) == null) {
+                    Horse createdHorse = horseDAO.create(horse);
+                    transaction.commit();
+
+                    messages.add(messageResource.get("dashboard.horse.create.success"));
+                    content.insertSessionAttribute(REQUEST_MESSAGES, messages);
+
+                    LOGGER.log(Level.DEBUG, "Created horse: " + createdHorse);
+                } else {
+                    String errorMessage = messageResource.get("dashboard.horse.create.duplicated");
+                    transaction.rollback();
+
+                    messages.add(errorMessage);
+                    content.insertSessionAttribute(REQUEST_ERRORS, messages);
+
+                    throw new ReceiverException(errorMessage);
+                }
             } catch (DAOException e) {
+                transaction.rollback();
+
                 messages.add(messageResource.get("dashboard.horse.create.fail"));
                 content.insertSessionAttribute(REQUEST_ERRORS, messages);
 
@@ -79,6 +100,8 @@ public class HorseReceiverImpl extends AbstractReceiver implements HorseReceiver
                 }
 
                 throw new ReceiverException("Database Error: " + e.getMessage(), e);
+            } finally {
+                transaction.endTransaction();
             }
         } else {
             content.insertSessionAttribute(REQUEST_ERRORS, validator.getErrors());
@@ -100,9 +123,8 @@ public class HorseReceiverImpl extends AbstractReceiver implements HorseReceiver
      */
     @Override
     public void ajaxObtainHorsesList(RequestContent content) throws ReceiverException {
+        Locale locale = (Locale) content.findSessionAttribute(SESSION_LOCALE);
         MessageWrapper messages = new MessageWrapper();
-
-        Locale locale = (Locale)content.findSessionAttribute(SESSION_LOCALE);
         MessageConfig messageResource = new MessageConfig(locale);
 
         try (HorseDAOImpl horseDAO = new HorseDAOImpl(false)) {
@@ -128,13 +150,12 @@ public class HorseReceiverImpl extends AbstractReceiver implements HorseReceiver
      */
     @Override
     public void removeHorse(RequestContent content) throws ReceiverException {
-        String idValue = content.findParameter(RequestFieldConfig.Horse.ID_FIELD);
-
-        Locale locale = (Locale)content.findSessionAttribute(SESSION_LOCALE);
-        HorseValidator validator = new HorseValidator(locale);
-
-        MessageWrapper messages = new MessageWrapper();
+        Locale locale = (Locale) content.findSessionAttribute(SESSION_LOCALE);
         MessageConfig messageResource = new MessageConfig(locale);
+        MessageWrapper messages = new MessageWrapper();
+
+        String idValue = content.findParameter(RequestFieldConfig.Horse.ID_FIELD);
+        HorseValidator validator = new HorseValidator(locale);
 
         if (validator.validateRemoveHorse(idValue)) {
             int id = Integer.parseInt(idValue);
@@ -142,16 +163,35 @@ public class HorseReceiverImpl extends AbstractReceiver implements HorseReceiver
             Horse suit = new Horse(id);
             LOGGER.log(Level.DEBUG, "Want remove horse: " + suit);
 
-            try (HorseDAOImpl horseDAO = new HorseDAOImpl(false)) {
-                horseDAO.remove(suit);
+            HorseDAOImpl horseDAO = new HorseDAOImpl(true);
 
-                messages.add(messageResource.get("dashboard.horse.remove.success"));
-                content.insertSessionAttribute(REQUEST_MESSAGES, messages);
+            TransactionManager transaction = new TransactionManager(horseDAO);
+            transaction.beginTransaction();
+
+            try {
+                if (horseDAO.find(id) != null) {
+                    horseDAO.remove(suit);
+                    transaction.commit();
+
+                    messages.add(messageResource.get("dashboard.horse.remove.success"));
+                    content.insertSessionAttribute(REQUEST_MESSAGES, messages);
+                } else {
+                    transaction.rollback();
+
+                    messages.add(messageResource.get("dashboard.horse.remove.undefined"));
+                    content.insertSessionAttribute(REQUEST_ERRORS, messages);
+
+                    throw new ReceiverException(messageResource.get("dashboard.horse.undefined"));
+                }
             } catch (DAOException e) {
-                messages.add(messageResource.get("dashboard.horse.remove.fail"));
+                transaction.rollback();
+
+                messages.add(messageResource.get("dashboard.horse.fail"));
                 content.insertSessionAttribute(REQUEST_ERRORS, messages);
 
                 throw new ReceiverException("Database Error: " + e.getMessage(), e);
+            } finally {
+                transaction.endTransaction();
             }
         } else {
             content.insertSessionAttribute(REQUEST_ERRORS, validator.getErrors());
@@ -166,24 +206,25 @@ public class HorseReceiverImpl extends AbstractReceiver implements HorseReceiver
      */
     @Override
     public void updateHorse(RequestContent content) throws ReceiverException {
+        Locale locale = (Locale) content.findSessionAttribute(SESSION_LOCALE);
+
+        MessageWrapper messages = new MessageWrapper();
+        MessageConfig messageResource = new MessageConfig(locale);
+
         String idAttr = content.findParameter(RequestFieldConfig.Horse.ID_FIELD);
         String name = content.findParameter(RequestFieldConfig.Horse.NAME_FIELD);
         String genderAttr = content.findParameter(RequestFieldConfig.Horse.GENDER_FIELD);
         String ageAttr = content.findParameter(RequestFieldConfig.Horse.AGE_FIELD);
         String suitAttr = content.findParameter(RequestFieldConfig.Horse.SUIT_FIELD);
 
-        Locale locale = (Locale)content.findSessionAttribute(SESSION_LOCALE);
         HorseValidator validator = new HorseValidator(locale);
-
-        MessageWrapper messages = new MessageWrapper();
-        MessageConfig messageResource = new MessageConfig(locale);
 
         if (validator.validateUpdateHorse(idAttr, name, genderAttr, ageAttr, suitAttr)) {
             int id = Integer.parseInt(idAttr);
             int suitId = Integer.parseInt(suitAttr);
 
             byte age = Byte.parseByte(ageAttr);
-            boolean gender = "male".equals(genderAttr);
+            boolean gender = HorseGenderType.isMale(genderAttr);
 
             Horse horse = new Horse(id);
 
@@ -194,12 +235,30 @@ public class HorseReceiverImpl extends AbstractReceiver implements HorseReceiver
 
             LOGGER.log(Level.DEBUG, "Want update horse: " + horse);
 
-            try (HorseDAOImpl horseDAO = new HorseDAOImpl(false)) {
-                horseDAO.update(horse);
+            HorseDAOImpl horseDAO = new HorseDAOImpl(true);
 
-                messages.add(messageResource.get("dashboard.horse.update.success"));
-                content.insertSessionAttribute(REQUEST_MESSAGES, messages);
+            TransactionManager transaction = new TransactionManager(horseDAO);
+            transaction.beginTransaction();
+
+            try {
+                if (horseDAO.find(id) != null) {
+                    horseDAO.update(horse);
+                    transaction.commit();
+
+                    messages.add(messageResource.get("dashboard.horse.update.success"));
+                    content.insertSessionAttribute(REQUEST_MESSAGES, messages);
+                } else {
+                    transaction.rollback();
+
+                    messages.add(messageResource.get("dashboard.horse.undefined"));
+                    content.insertSessionAttribute(REQUEST_ERRORS, messages);
+
+                    throw new ReceiverException(messageResource.get("dashboard.horse.undefined"));
+                }
+
             } catch (DAOException e) {
+                transaction.rollback();
+
                 messages.add(messageResource.get("dashboard.horse.update.fail"));
                 content.insertSessionAttribute(REQUEST_ERRORS, messages);
 
@@ -208,6 +267,8 @@ public class HorseReceiverImpl extends AbstractReceiver implements HorseReceiver
                 }
 
                 throw new ReceiverException("Database Error: " + e.getMessage(), e);
+            } finally {
+                transaction.endTransaction();
             }
         } else {
             content.insertSessionAttribute(REQUEST_ERRORS, validator.getErrors());
